[issue-7] extract Vector and Point into new noray::math module
